move id2Network logic
